pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lineagescope"
version = "0.2.0"
description = "Universal static analyzer for data pipelines"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Kiran", email = "kiran.m.narayana@gmail.com" }]
keywords = [
    "data-engineering",
    "lineage",
    "lineagescope",
    "dbt",
    "airflow",
    "static-analysis",
]
dependencies = [
    "sqlglot>=26.0",
    "networkx>=3.0",
    "typer>=0.12",
    "rich>=13.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "gitpython>=3.1",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.8",
    "pre-commit>=3.0",
    "Pillow>=10.0",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "pymdown-extensions>=10.0",
]

[project.urls]
Documentation = "https://kirannarayanak.github.io/lineagescope/"
Repository = "https://github.com/kirannarayanak/lineagescope"
Changelog = "https://github.com/kirannarayanak/lineagescope/blob/main/CHANGELOG.md"

[project.scripts]
lineagescope = "lineagescope.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["lineagescope"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
packages = ["lineagescope"]
strict = false
warn_return_any = true
warn_unused_ignores = true
